requirements.txt
aiofiles==23.2.1
aiogram==2.23.1       
aiohttp==3.8.6        
aiosignal==1.3.1      
annotated-types==0.6.0
anyio==4.3.0          
asgiref==3.7.2        
async-timeout==4.0.3  
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
click-log==0.4.0
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
deprecation==2.1.0
Django==3.2
django-bootstrap4==23.2
django-crispy-forms==1.14.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.2.0
Flask==2.3.2
fonttools==4.44.0
frozenlist==1.4.1
gitdb==4.0.10
GitPython==3.1.32
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gotrue==2.4.1
gspread==5.10.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.2.0
kiwisolver==1.4.5
load-dotenv==0.1.0
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.8.1
mdurl==0.1.2
more-itertools==10.0.0
multidict==6.0.5
nh3==0.2.15
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==24.0
Pillow==10.0.1
pkginfo==1.9.6
postgrest==0.16.1
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyDrive==1.3.1
Pygments==2.15.1
pyloco==0.0.139
pyparsing==3.1.1
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
raspi-gpio==0.1.9
readme-renderer==40.0
realtime==1.0.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
schedule==1.2.0
SecretStorage==3.3.3
semver==2.13.0
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tomlkit==0.12.1
tqdm==4.65.0
twine==3.8.0
typing==3.7.4.3
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.16
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.4
zipp==3.16.2
